+++ NEWS.md
@@ -1,6 +1,9 @@
 ## formatters 0.5.2.9002
   * Specified minimal version of package dependencies.
 
+### Miscellaneous
+ * Decimal alignment now throws informative error if scientific notation is used.
+
 ## formatters 0.5.2
  * `paginate_to_mpfs` can handle single column tables and listings.
  * Added decimal alignment support with `decimal`, `dec_right`, and `dec_left`.